Adjust the log info
