+++ test/chunking-form/samples/circular-chunks-warning-caused-by-only-explicit-manual-chunks/_expected/amd/generated-ac.js
@@ -1,10 +1,10 @@
-define(['exports', './main'], (function (exports, main) { 'use strict';
+define(['exports', './generated-b'], (function (exports, b) { 'use strict';
 
 	const c = 'c';
 	console.log(c);
 
 	const a = 'a';
-	console.log(a + main.b);
+	console.log(a + b.b);
 
 	exports.a = a;
 	exports.c = c;
