+++ test/chunking-form/samples/circular-chunks-warning-caused-by-only-explicit-manual-chunks/_expected/amd/generated-b.js
@@ -1,0 +1,8 @@
+define(['exports', './generated-ac'], (function (exports, ac) { 'use strict';
+
+	const b = 'b';
+	console.log(b + ac.c);
+
+	exports.b = b;
+
+}));
