+++ test/chunking-form/samples/circular-chunks-warning-caused-by-only-explicit-manual-chunks/_expected/amd/main.js
@@ -1,10 +1,5 @@
-define(['exports', './generated-ac', './main'], (function (exports, ac, main) { 'use strict';
-
-	const b = 'b';
-	console.log(b + ac.c);
+define(['./generated-ac', './generated-b'], (function (ac, b) { 'use strict';
 
 	console.log(ac.a);
-
-	exports.b = b;
 
 }));
