+++ test/chunking-form/samples/circular-chunks-warning-caused-by-only-explicit-manual-chunks/_expected/es/generated-ac.js
@@ -1,4 +1,4 @@
-import { b } from './main.js';
+import { b } from './generated-b.js';
 
 const c = 'c';
 console.log(c);
