+++ test/chunking-form/samples/circular-chunks-warning-caused-by-only-explicit-manual-chunks/_expected/es/generated-b.js
@@ -1,0 +1,6 @@
+import { c } from './generated-ac.js';
+
+const b = 'b';
+console.log(b + c);
+
+export { b };
