+++ test/chunking-form/samples/circular-chunks-warning-caused-by-only-explicit-manual-chunks/_expected/system/generated-ac.js
@@ -1,4 +1,4 @@
-System.register(['./main.js'], (function (exports) {
+System.register(['./generated-b.js'], (function (exports) {
 	'use strict';
 	var b;
 	return {
