+++ test/chunking-form/samples/circular-chunks-warning-caused-by-only-explicit-manual-chunks/_expected/system/generated-b.js
@@ -1,0 +1,15 @@
+System.register(['./generated-ac.js'], (function (exports) {
+	'use strict';
+	var c;
+	return {
+		setters: [function (module) {
+			c = module.c;
+		}],
+		execute: (function () {
+
+			const b = exports("b", 'b');
+			console.log(b + c);
+
+		})
+	};
+}));
